remove PlansController#show
This action is being replaced by the PagesController#planner action. There is no reason to be viewing a Plan that is not the active plan as the whole context of plan-related actions on the website assume the user is working with the active plan. Note that adding a PlannedCourse to a non-active plan is not actually blocked, just not possible to do via the current UI. This should be sufficient.